backend/requirements.txt
upstash-redis
python-dotenv
crewai
crewai[google-genai]
crewai-tools
fastapi
uvicorn
pydantic
langchain
langchain_core
langchain_community
langchain_chroma
langchain-groq
beautifulsoup4
ddgs
requests
mcp
starlette
pandas
langchain-cohere
httpx